Add player launcher plugin
Current implementation from suggested practice in that onCommand still
executes conditionally on whether label is equal to the command "launch"
or its alias "lch", whereas no such check is employed by the latter.

The working hypothesis is that because the string used when registering
the command and declaring in the plugin.yml differs from the spelling in
expected usage, it is imperative that the input from the label is
checked to be equal to the expected usage, rather than the name of the
string used when internally registering the command. Furthermore, the
default expected usage appears to be equivalent to how the command is
internally registered.